# specs/001-batfish-snapshot-topology/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Batfish Visualization & Verification API
  description: |
    REST API for Batfish snapshot creation, topology visualization, and configuration verification.

    This API integrates with Batfish v2025.07.07 container via pybatfish v2025.07.07.

    **Features**:
    - Snapshot creation from network configuration files
    - Layer 3 topology query and visualization
    - Configuration verification queries (reachability, ACL, routing)
  version: 1.0.0
  contact:
    name: Batfish Vis Team
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://batfish-vis.example.com/api
    description: Production server
paths:
...
  /snapshots:
    get:
      summary: List snapshots
      description: Get list of all Batfish snapshots
      operationId: listSnapshots
      tags:
        - Snapshots
      responses:
        '200':
          description: List of snapshots
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items:
                      $ref: '#/components/schemas/Snapshot'

    post:
      summary: Create snapshot
      description: Upload network configuration files and create Batfish snapshot
      operationId: createSnapshot
      tags:
        - Snapshots
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                snapshotName:
                  type: string
                  description: Unique snapshot name
                  example: "snapshot-2025-11-21-14-30"
                networkName:
                  type: string
                  description: Network name (logical grouping)
                  example: "production-network"
                  default: "default"
                configFiles:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Network configuration files
              required:
                - snapshotName
                - configFiles
      responses:
        '201':
          description: Snapshot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '400':
          description: Invalid request (missing files, invalid name, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Snapshot with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Batfish container unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /snapshots/{snapshotName}:
    get:
      summary: Get snapshot details
      description: Get detailed information about a specific snapshot
      operationId: getSnapshot
      tags:
        - Snapshots
      parameters:
        - name: snapshotName
          in: path
          required: true
          schema:
            type: string
          description: Snapshot name
      responses:
        '200':
          description: Snapshot details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snapshot'
        '404':
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete snapshot
      description: Delete a Batfish snapshot
      operationId: deleteSnapshot
      tags:
        - Snapshots
      parameters:
        - name: snapshotName
          in: path
          required: true
          schema:
            type: string
          description: Snapshot name
      responses:
        '204':
          description: Snapshot deleted successfully
        '404':
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Snapshot:
      type: object
      properties:
        name:
          type: string
          example: "snapshot-2025-11-21-14-30"
        network:
          type: string
          example: "production-network"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-21T14:30:15Z"
        status:
          type: string
          enum: [COMPLETE, FAILED, IN_PROGRESS]
          example: "COMPLETE"
        configFileCount:
          type: integer
          example: 15
        deviceCount:
          type: integer
          example: 12
        batfishVersion:
          type: string
          example: "v2025.07.07"
        parseErrors:
          type: array
          items:
            type: object
            properties:
              fileName:
                type: string
              errorMessage:
                type: string
              lineNumber:
                type: integer
                nullable: true
      required:
        - name
        - network
        - createdAt
        - status
        - configFileCount
        - deviceCount
        - batfishVersion
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Snapshot not found"
        message:
          type: string
          example: "The snapshot 'invalid-snapshot' does not exist"
        code:
          type: string
          example: "SNAPSHOT_NOT_FOUND"
      required:
        - error
        - message
...